Vite F1/F2: load from dist, extract shared avatarRenderer, fix assets + modelview CSS
- main.js / framework.js load the renderer pages from the Vite build output
  (dist/), preserving the shared file:// origin for the localStorage handoff
  (Invariant 1). package.json: add `build`, make `start` = `vite build && electron .`.
- Extract the triplicated three.js scene/load/rig/apply/animate/record logic into
  src/render/avatarRenderer.js (data-source-agnostic, with the cached new Function
  binding) + src/data/ipcSource.js. render/render.js is now a ~60-line wrapper.
  The discrete render path is validated working (incl. the FBX binding path);
  mocaprender still uses its own copy until it's migrated next.
- vite.config: copy built-in models/ into dist (models.json path/picBg are runtime
  strings Vite can't see).
- modelview.html: !important on body font-family + background so they beat mdui's
  bundled CSS (Vite injects it after the inline <style>) — restores custom fonts
  and the acrylic glass.
- avatarRenderer drop handler: e.preventDefault() so dropping a background image no
  longer makes Electron open/navigate to the file.
- docs/REFACTOR.md: record the build gotchas (CSS order, runtime-string assets, drop).
